+++ src/core/b2connect.py
@@ -1,20 +1,13 @@
 from b2sdk.exception import *
 from b2sdk.v2 import *
 import os
-from dotenv import load_dotenv
 
-load_dotenv()
-
-cache_folder = os.getenv('CACHE_FOLDER')
-os.makedirs(cache_folder, exist_ok=True)
-
-b2_app_id = os.getenv('B2_KEY_ID')
-b2_app_key = os.getenv('B2_APP_KEY')
+from core.config import cache_folder, b2_app_id, b2_app_key, b2_bucket_name
 
 info = InMemoryAccountInfo()
 b2_api = B2Api(info)
 b2_api.authorize_account("production", b2_app_id, b2_app_key)
-b2_bucket = b2_api.get_bucket_by_name(os.getenv('B2_BUCKET_NAME'))
+b2_bucket = b2_api.get_bucket_by_name(b2_bucket_name)
 
 
 def b2_file_upload(filepath):
