+++ src/routes/interface.py
@@ -1,0 +1,7 @@
+from flask import render_template
+from core.config import app
+
+
+@app.route("/")
+def index():
+    return "ok"
